Deactivate memcached support
Turn back to development default (phpsession). Memcache requires to install the memcached daemon and a PHP library (Memcache); the process is not documented in the installation instructions and I wasn't able to even find the correct PHP library and/or install (there were always conflicts of dependencies). I think that for development purposes, it's better to use the default that requires less dependencies.